requirements.txt
awscli==1.27.86
awscli-local==0.20
boto3==1.26.86
botocore==1.29.86
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.4
docutils==0.16
idna==3.4
jmespath==1.0.1
localstack-client==1.39
psycopg2==2.9.5
pyasn1==0.4.8
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==5.4.1
requests==2.28.2
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
